third_party/WebKit/LayoutTests/webaudio/mediastreamaudiodestinationnode-expected.txt

Basic tests for MediaStreamAudioDestinationNode API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS mediaStreamDestination.__proto__.__proto__ is AudioNode.prototype
PASS Destination AudioNode has one input.
PASS successfullyParsed is true

TEST COMPLETE
